Abstracting jwt/helper functions from server code
